requirements.txt
Jinja2==3.1.4
python-slugify==8.0.4
